Changed the project structure
