--- veceditor/App.axaml.cs
@@ -1,24 +1,0 @@
-using Avalonia;
-using Avalonia.Controls.ApplicationLifetimes;
-using Avalonia.Markup.Xaml;
-
-namespace veceditor
-{
-   public partial class App : Application
-   {
-      public override void Initialize()
-      {
-         AvaloniaXamlLoader.Load(this);
-      }
-
-      public override void OnFrameworkInitializationCompleted()
-      {
-         if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
-         {
-            desktop.MainWindow = new MainWindow();
-         }
-
-         base.OnFrameworkInitializationCompleted();
-      }
-   }
-}
